requirements.txt
pillow
opencv-python
numpy
paddlepaddle
pyclipper
Shapely